Some bugs were solved.

+++ app.js
@@ -26,7 +26,7 @@ app.use((req, res, next) => {
     throw error
 })
 app.use((error, req, res, next) => {
-    if (req.file) {
+    if (req.file && req.route.path !=='/signup') {
         fs.unlink(req.file.path, (err) => {
             console.log(err);
         })
